requirements.txt
python-docx
bs4
pillow